refactor: separate concerns in src/main.py
Extract bot business logic into BotService and all routes into a
dedicated APIRouter, leaving main.py as a pure wiring/bootstrap module.

Closes #31
